Cache classloaders when loading from resource.
Note that this seems to be a good idea to avoid the overhead of loading
classloaders over and over again. At the same time, it may be bad
because we hold on to classloaders indefinitely (well...). Also, there
is currently no way to test this since we don't have unit tests for the
inner workings of require.